requirements.txt
flask 
mysql.connector
datetime